Split TSM.py into multiple files
